requirements.txt
dlib==19.21.1
matplotlib==3.3.0
Keras==2.4.3
imutils==0.5.4
pandas==1.1.0
mtcnn==0.1.0
numpy==1.17.4
scikit_image==0.18.1
opencv_python==4.5.1.48
tqdm==4.54.1
face_recognition==1.3.0
scipy==1.5.4
tensorflow==2.3.0
Augmentor==0.2.8
Pillow==8.1.2
scikit_learn==0.24.1
skimage==0.0
